Task/improve string format performance

+++ examples/Collection/Arrays/main.cpp
@@ -1,7 +1,33 @@
+/**
+ * Copyright 2017 Food Tiny Project. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * Redistributions of source code must retain the above copyright
+ * notice, this list of conditions and the following disclaimer.
+ * Redistributions in binary form must reproduce the above
+ * copyright notice, this list of conditions and the following disclaimer
+ * in the documentation and/or other materials provided with the
+ * distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
 #include "../../../library.hpp"
 
 void ArraysSort() {
-	long arrayLongNormal[5] = { 3, -2, 4, 0, -1 };
+	long long arrayLongNormal[5] = { 3, -2, 4, 0, -1 };
 	Arrays::sort(arrayLongNormal, 0, 4);
 	
 	Array<long> arrayLongArray = { 3, -2, 4, 0, -1 };
@@ -9,14 +35,14 @@ void ArraysSort() {
 }
 
 void ArraysSearch() {
-	long arrayLongNormal[5] = { 3, -2, 4, 0, -1 };
+	long long arrayLongNormal[5] = { 3, -2, 4, 0, -1 };
 	Arrays::binarySearch(arrayLongNormal, 4, 5);
 	
 	Array<long> arrayLongArray = { 3, -2, 4, 0, -1 };
 	Arrays::binarySearch(arrayLongArray, 4L, 5);
 }
 
-int main() {
+long int main() {
 	ArraysSort();
 	ArraysSearch();
 	return 0;
